refactor: sole responsibility for the favorites component
