+++ src/components/common/Favorite.vue
@@ -13,6 +13,7 @@
       hover:text-default-secondary
       cursor-pointer
     "
+    @click="toggleFavoritePokemon"
   >
     <StarIcon
       class="fill-current hover:text-default-secondary"
@@ -22,18 +23,39 @@
 </template>
 
 <script lang="ts">
-import Vue from "vue";
+import Vue, { PropOptions } from "vue";
+import { mapState, mapMutations } from "vuex";
 import StarIcon from "@/assets/img/star.svg";
+import { Pokemon } from "@/interfaces/pokemon";
 
 export default Vue.extend({
   components: {
     StarIcon,
   },
   props: {
-    isFavorite: {
-      type: Boolean,
-      default: false,
+    pokemon: {
+      type: Object,
+      default: () => {},
+    } as PropOptions<Pokemon>,
+  },
+  computed: {
+    isFavorite(): Boolean {
+      const pokemon = this.favoritePokemons.find(
+        (p: Pokemon) => p.name === this.pokemon.name
+      );
+      let isFavorite = this.favoritePokemons.includes(pokemon);
+      return isFavorite;
     },
+    ...mapState(["favoritePokemons"]),
+  },
+  methods: {
+    /**
+     * Method to mark or unmark favorite pokemon
+     */
+    toggleFavoritePokemon() {
+      this.setFavoritePokemons(this.pokemon);
+    },
+    ...mapMutations(["setFavoritePokemons"]),
   },
 });
 </script>
