+++ src/components/list/PokemonItem.vue
@@ -4,13 +4,12 @@
     @click.self="viewDetail"
   >
     <span class="capitalize"> {{ pokemon.name }} </span>
-    <Favorite @click.native="toggleFavoritePokemon" :isFavorite="isFavorite" />
+    <Favorite :pokemon="pokemon" />
   </div>
 </template>
 
 <script lang="ts">
 import Vue, { PropOptions } from "vue";
-import { mapMutations, mapState } from "vuex";
 import { Pokemon } from "@/interfaces/pokemon";
 import Favorite from "../common/Favorite.vue";
 
@@ -27,30 +26,13 @@ export default Vue.extend({
       required: true,
     } as PropOptions<Pokemon>,
   },
-  computed: {
-    isFavorite(): Boolean {
-      const pokemon = this.favoritePokemons.find(
-        (p: Pokemon) => p.name === this.pokemon.name
-      );
-      let isFavorite = this.favoritePokemons.includes(pokemon);
-      return isFavorite;
-    },
-    ...mapState(["favoritePokemons"]),
-  },
   methods: {
-    /**
-     * Method to mark or unmark favorite pokemon
-     */
-    toggleFavoritePokemon() {
-      this.setFavoritePokemons(this.pokemon);
-    },
     /**
      * Method to see pokemon detail
      */
     viewDetail() {
       this.$emit("pokemonName", this.pokemon.name);
     },
-    ...mapMutations(["setFavoritePokemons"]),
   },
 });
 </script>
